root stack created in app.js
