+++ package.json
@@ -13,6 +13,7 @@
     "expo": "^30.0.1",
     "react": "16.3.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz",
+    "react-navigation": "^2.16.0",
     "yelp-api": "^1.0.3",
     "yelp-fusion": "^2.2.1"
   }
