+++ screens/HomeScreen.js
@@ -1,0 +1,8 @@
+import React, { Components } from "react";
+import { Text, View, Button } from "react-native";
+
+export default class HomeScreen extends React.Component {
+  render() {
+    return <View />;
+  }
+}
